+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## Version 0.2.0
+
+- Add more parameter and status checks
+- Add InSpec support by properly handling failing commands
+- Add Windows support
+- Refactor code
+
 ## Version 0.1.1
 
 - Add support for passing in an instance-id instead of IP/DNS